src/main/java/com/summer/intern/employees/Services/EmployeeService.java
package com.summer.intern.employees.Services;

import com.summer.intern.employees.Models.Employee;
import com.summer.intern.employees.Repositories.EmployeeRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class EmployeeService {

    private EmployeeRepository employeeRepository;
    @Autowired
    public EmployeeService(EmployeeRepository employeeRepository) {
        this.employeeRepository=employeeRepository;
    }

    public List<Employee> getEmployees() {
        return employeeRepository.findAll();
    }

    public boolean exist(String id) {
        return employeeRepository.existsById(id);
    }

    public Employee getEmployee(String id) {
        return employeeRepository.findById(id).get();
    }


    public void addEmployee(Employee employee) {
        employeeRepository.save(employee);
    }

    public void updateEmployee(Employee employee) {
        employeeRepository.save(employee);
    }

    public void deleteEmployee(String id) {
        employeeRepository.deleteById(id);
    }
}
